fold: DECL_VALUE_EXPR isn't simple [PR120400]
This PR noted that fold_truth_andor was wrongly changing && to & where the
RHS is a VAR_DECL with DECL_VALUE_EXPR; we can't assume that such can be
evaluated unconditionally.

To be more precise we could recurse into DECL_VALUE_EXPR, but that doesn't
seem worth bothering with since typical uses involve a COMPONENT_REF, which
is not simple.

	PR c++/120400

gcc/ChangeLog:

	* fold-const.cc (simple_operand_p): False for vars with
	DECL_VALUE_EXPR.