c++, coroutines: Delete now unused code for parm guards.
Since r16-775-g18df4a10bc9694 we use nested cleanups to
handle parameter copy destructors in the ramp (and pass
a list of cleanups required to the actor which will only
be invoked if the parameter copies were all correctly
built - and therefore does not need to guard destructors
either.

This deletes the provisions for frame parameter copy
destructor guards.

gcc/cp/ChangeLog:

	* coroutines.cc (analyze_fn_parms): No longer
	create a parameter copy guard var.
	* coroutines.h (struct param_info): Remove the
	entry for the parameter copy destructor guard.

Signed-off-by: Iain Sandoe <iain@sandoe.co.uk>