ada: Fix alignment violation for mix of aligned and misaligned composite types
This happens when the chain of initialization procedures is called on the
subcomponents and causes the creation of temporaries along the way out of
alignment considerations.  Now these temporaries are not necessary in the
context and were not created until recently, so this gets rid of them.

gcc/ada/ChangeLog:

	* gcc-interface/trans.cc (addressable_p): Add COMPG third parameter.
	<COMPONENT_REF>: Do not return true out of alignment considerations
	for non-strict-alignment targets if COMPG is set.
	(Call_to_gnu): Pass true as COMPG in the call to the addressable_p
	predicate if the called subprogram is an initialization procedure.