libstdc++: Fix test failures for 32-bit AIX
With -maix32 (the default) we only have 16-bit wchar_t so these tests
fail. The debug.cc one is because we use -fwide-exec-charset=UTF-32BE
which tries to encode each wide character as four bytes in a 2-byte
wchar_t. The format.cc one is because the clown face character can't be
encoded in a single 16-bit wchar_t.

libstdc++-v3/ChangeLog:

	* testsuite/std/format/debug.cc: Disable for targets with 16-bit
	wchar_t.
	* testsuite/std/format/functions/format.cc: Use -DUNICODE for
	targets with 32-bit wchar_t.
	(test_unicode) [UNICODE]: Only run checks when UNICODE is
	defined.