Do not recompute profile when entry block has afdo count of 0
With normal profile feedback checking entry block count to be non-zero is quite
reliable check for presence of non-0 profile in the body since the function
body can only be executed if the entry block was executed.  With autofdo this
is not true, since the entry block may just execute too few times to be
recorded.  As a consequence we currently drop AFDO profile quite often.  This
patch fixes it.

gcc/ChangeLog:

	* predict.cc (rebuild_frequencies): look harder for presence
	of profile feedback.