c++, coroutines: Fix identification of coroutine ramps [PR120453].
The existing implementation, incorrectly, tried to use DECL_RAMP_FN
in check_return_expr to determine if we are handling a ramp func.
However, that query is only set for the resume/destroy functions.

Replace the use of DECL_RAMP_FN with a new query.

	PR c++/120453

gcc/cp/ChangeLog:

	* cp-tree.h (DECL_RAMP_P): New.
	* typeck.cc (check_return_expr): Use DECL_RAMP_P instead
	of DECL_RAMP_FN.

gcc/testsuite/ChangeLog:

	* g++.dg/coroutines/pr120453.C: New test.

Signed-off-by: Iain Sandoe <iain@sandoe.co.uk>