c: fix ICE for mutually recursive structures [PR120381]
For invalid nesting of a structure definition in a definition
of itself or when using a rather obscure construction using statement
expressions, we can create mutually recursive pairs of non-identical
but compatible structure types.  This can lead to invalid composite
types and an ICE.  If we detect recursion even for swapped pairs
when forming composite types, this is avoided.

	PR c/120381

gcc/c/ChangeLog:
	* c-typeck.cc (composite_type_internal): Stop recursion for
	swapped pairs.

gcc/testsuite/ChangeLog:
	* gcc.dg/pr120381.c: New test.
	* gcc.dg/gnu23-tag-composite-6.c: New test.