Have TODO_verify_* not set by any pass
This is a follow up to the patch set starting at https://gcc.gnu.org/pipermail/gcc-patches/2014-April/386650.html.

Currently TODO_verify_{il,all} is set by a few passes as TODOs afterwards but
we don't need to do that any more. Those were mostly removed back in
https://gcc.gnu.org/pipermail/gcc-patches/2014-May/387647.html

And now since TODO_verify_all is no longer used, remove it too.
Removes last_verified from the function structure since it is always 0 before the
call to execute_todo that has TODO_verify_all set on it.

Also this removes the ability for good to use TODO_verify_il from the passes and reserves
bit 31 of the TODO for the verfification and internally only the passes.cc code.

gcc/ChangeLog:

	* function.h (struct function): Remove last_verified.
	* gimple-harden-conditionals.cc (pass_data_harden_compares): Remove
	TODO_verify_il.
	(pass_data_harden_conditional_branches): Likewise.
	* gimple-harden-control-flow.cc (pass_harden_control_flow_redundancy::execute):
	Don't return TODO_verify_il.
	* ipa-strub.cc (pass_data_ipa_strub): Remove TODO_verify_il.
	* passes.cc (TODO_verify_il): Define.
	(execute_function_todo): Don't use or set last_verified.
	(clear_last_verified): Remove.
	(execute_one_ipa_transform_pass): Update comment before execute_todo.
	Assert that none of the todos have TODO_verify_il set on it.
	(execute_one_pass): Don't call clear_last_verified on all functions.
	Assert that none of the todos have TODO_verify_il set on it.
	* tree-inline.cc (initialize_cfun): Don't copy last_verified.
	* tree-pass.h (TODO_verify_all): Remove.
	* tree-vrp.cc (pass_data_early_vrp): Remove TODO_verify_all.
	(pass_data_fast_vrp): Likewise.

Signed-off-by: Andrew Pinski <quic_apinski@quicinc.com>