diagnostics: consolidate calls to colorizer::set_named_color
I noticed whilst working on another issue that in diagnostic-show-locus
within the quoted source lines and the annotation underlines that when
we're showing highlight-{a,b} that we emit
  start-colorization-code, character, end-colorization-code
per *character*, rather than just when the colorization changes.

This was due to me failing to implement consolidation of such changes in
colorizer::set_named_color in r15-2015-g7d73c01ce6d1ab.

Fixed thusly, simplifying the output.

I manually inspected all of the changed testcases in a terminal with
color enabled and verified that the output is visually identical to
before.

gcc/ChangeLog:
	* diagnostic-show-locus.cc (colorizer::m_current_named_color): New
	field.
	(colorizer::set_named_color): Use it to consolidate repeated calls
	to the same color.

gcc/testsuite/ChangeLog:
	* g++.dg/diagnostic/bad-binary-ops-highlight-colors.C: Update
	expected multiline output for quoted source and underlines to
	reflect emitting color codes when changes happen, rather than
	per character.
	* g++.dg/diagnostic/long-short-colorization.C: Likewise.
	* g++.dg/plugin/show-template-tree-color-labels.C: Likewise.
	* gcc.dg/bad-binary-ops-highlight-colors.c: Likewise.
	* gcc.dg/format/colors.c: Likewise.

Signed-off-by: David Malcolm <dmalcolm@redhat.com>